Move util method. No functional changes.
