+++ packages/devtools/lib/src/framework/framework.dart
@@ -20,6 +20,7 @@ import '../ui/custom.dart';
 import '../ui/elements.dart';
 import '../ui/primer.dart';
 import '../ui/ui_utils.dart';
+import '../url_utils.dart';
 import '../utils.dart';
 import 'framework_core.dart';
