+++ packages/devtools/lib/src/utils.dart
@@ -171,17 +171,6 @@ String getSimpleStackFrameName(String name) {
   return name;
 }
 
-/// Returns a trimmed vm service uri without any trailing characters.
-///
-/// For example, given a [value] of http://127.0.0.1:60667/72K34Xmq0X0=/#/vm,
-/// this method will return the URI http://127.0.0.1:60667/72K34Xmq0X0=/.
-Uri getTrimmedUri(String value) {
-  final uri = Uri.parse(value.trim());
-  return uri
-      .removeFragment()
-      .replace(path: uri.path.endsWith('/') ? uri.path : '${uri.path}/');
-}
-
 class Property<T> {
   Property(this._value);
