+++ packages/devtools/test/url_utils_test.dart
@@ -23,6 +23,29 @@ void main() {
         equals('package:flutter_web/lib/src/widgets/binding.dart'),
       );
     });
+
+    test('getTrimmedUri', () {
+      expect(
+        getTrimmedUri('http://127.0.0.1:60667/72K34Xmq0X0=/#/vm').toString(),
+        equals('http://127.0.0.1:60667/72K34Xmq0X0=/'),
+      );
+      expect(
+        getTrimmedUri('http://127.0.0.1:60667/72K34Xmq0X0=').toString(),
+        equals('http://127.0.0.1:60667/72K34Xmq0X0=/'),
+      );
+      expect(
+        getTrimmedUri('http://127.0.0.1:60667/72K34Xmq0X0=/   ').toString(),
+        equals('http://127.0.0.1:60667/72K34Xmq0X0=/'),
+      );
+      expect(
+        getTrimmedUri('http://127.0.0.1:60667').toString(),
+        equals('http://127.0.0.1:60667/'),
+      );
+      expect(
+        getTrimmedUri('http://127.0.0.1:60667/').toString(),
+        equals('http://127.0.0.1:60667/'),
+      );
+    });
   });
 }
